Making Language dependent of the requested Domain.
multi_language? will return true if more than one published languages are found for the requested domain. Else it will return false.

Saving the domain_id to the session will be the first in order of all before filters.

Changing the language_id in the session can be only done if the language is assigned to the requested Domain.